card delete card added

+++ src/components/Card.js
@@ -1,11 +1,15 @@
 export default class Card {
-  constructor(name, link, likes, cardTemplate, handleCardClick, handleDeleteCard) {
-    this._name = name;
-    this._link = link;
-    this._likes = likes;
+  constructor(data, cardTemplate, handleCardClick, handleDeleteCard, handleLike, ownerId) {
+    this._name = data.name;
+    this._link = data.link;
+    this._likes = data.likes.length;
+    this._cardId = data._id;
+    this._pageOwnerId = ownerId;
+    this._ownerId = data.owner._id;
     this._cardBlock = cardTemplate.querySelector('.element').cloneNode(true);
     this._handleCardClick = handleCardClick;
     this._handleDeleteCard = handleDeleteCard;
+    this._handleLike = handleLike;
     this._cardImage = this._cardBlock.querySelector('.element__image');
     this._likeCounter = this._cardBlock.querySelector('.element__counter')
   }
@@ -13,8 +17,14 @@ export default class Card {
   createCard() {
     this._cardImage.src = this._link;
     this._cardImage.alt = this._name;
+    this._cardBlock.id = this._cardId;
     this._likeCounter.textContent = this._likes;
     this._cardBlock.querySelector('.element__title').textContent = this._name;
+    console.log(this._ownerId, this._pageOwnerId);
+
+    if (this._pageOwnerId !== this._ownerId) {
+      this._cardBlock.querySelector('.element__trash').remove();
+    }
 
     this._setEventListeners();
 
@@ -22,12 +32,15 @@ export default class Card {
   }
 
   _setEventListeners() {
+    const trashIcon = this._cardBlock.querySelector('.element__trash');
     this._cardBlock.querySelector('.element__like').addEventListener('click', (evt) => {
-      evt.target.classList.toggle('element__like_active');
+      this._handleLike(evt);
     });
-    this._cardBlock.querySelector('.element__trash').addEventListener('click', (evt) => {
-      this._handleDeleteCard(evt);
-    });
+    if (this._cardBlock.contains(trashIcon)) {
+      this._cardBlock.querySelector('.element__trash').addEventListener('click', (evt) => {
+        this._handleDeleteCard(evt);
+      });
+    }
     this._cardImage.addEventListener('click', () => this._handleCardClick(this._name, this._link));
   }
 }
